+++ Models/Role.cs
@@ -13,7 +13,7 @@ public class Role : Entity
     /// <summary>
     /// The users that have this role.
     /// </summary>
-    public required IEnumerable<ulong> UsersWithRoleIdentifiers { get; set; }
+    public required IEnumerable<uint> UsersWithRoleIdentifiers { get; set; }
 
     /// <summary>
     /// The users that have this role.
